# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoX Hunter AI Agent API",
    "version": "1.0.0",
    "description": "API for external AI agents to fetch leads, generate prototypes, and manage CRM outreach."
  },
  "servers": [
    {
      "url": "https://mox.infni-t.online"
    }
  ],
  "paths": {
...
    "/api/mcp/leads/{id}": {
      "get": {
        "operationId": "getLeadById",
        "summary": "Fetch a single lead by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the lead"
          }
        ],
        "responses": {
          "200": {
            "description": "Single lead details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateLead",
        "summary": "Update lead properties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the lead"
          }
        ],
        "responses": {
          "200": {
            "description": "Lead updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteLead",
        "summary": "Delete a lead from the CRM",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the lead"
          }
        ],
        "responses": {
          "200": {
            "description": "Lead deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "mo-x-api-key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}